requirements.txt
arabic-reshaper==2.1.1
beautifulsoup4==4.9.3
cycler==0.10.0
future==0.18.2
kiwisolver==1.3.1
lxml==4.6.2
matplotlib==3.3.3
numpy==1.19.5
Pillow==8.1.0
pyaes==1.6.1
pyasn1==0.4.8
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.1
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.1
Telethon==1.18.2
unicodecsv==0.14.1
urllib3==1.26.2
wordcloud==1.8.1
